pmd & lgtm fixes

+++ client/src/main/java/io/opencmw/client/DataSourcePublisher.java
@@ -1,5 +1,6 @@
 package io.opencmw.client;
 
+import static io.opencmw.OpenCmwProtocol.EMPTY_FRAME;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Objects.requireNonNull;
 
@@ -24,6 +25,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,11 +37,13 @@ import org.zeromq.ZMsg;
 
 import io.opencmw.EventStore;
 import io.opencmw.Filter;
+import io.opencmw.MimeType;
 import io.opencmw.OpenCmwProtocol.Command;
 import io.opencmw.QueryParameterParser;
 import io.opencmw.RingBufferEvent;
 import io.opencmw.client.cmwlight.CmwLightDataSource;
 import io.opencmw.client.rest.RestDataSource;
+import io.opencmw.domain.BinaryData;
 import io.opencmw.filter.EvtTypeFilter;
 import io.opencmw.filter.FilterRegistry;
 import io.opencmw.rbac.RbacProvider;
@@ -92,8 +96,7 @@ import io.opencmw.utils.SystemProperties;
 public class DataSourcePublisher implements Runnable, Closeable {
     public static final int MIN_FRAMES_INTERNAL_MSG = 3;
     private static final Logger LOGGER = LoggerFactory.getLogger(DataSourcePublisher.class);
-    private static final byte[] EMPTY_BYTE_ARRAY = new byte[0];
-    private static final ZFrame EMPTY_FRAME = new ZFrame(EMPTY_BYTE_ARRAY);
+    private static final ZFrame EMPTY_ZFRAME = new ZFrame(EMPTY_FRAME);
     private static final AtomicInteger INSTANCE_COUNT = new AtomicInteger();
     protected final long heartbeatInterval = SystemProperties.getValueIgnoreCase(HEARTBEAT, HEARTBEAT_DEFAULT); // [ms] time between to heartbeats in ms
     private final String inprocCtrl = "inproc://dsPublisher#" + INSTANCE_COUNT.incrementAndGet();
@@ -103,16 +106,17 @@ public class DataSourcePublisher implements Runnable, Closeable {
     private final AtomicBoolean running = new AtomicBoolean(false);
     private final AtomicInteger internalReqIdGenerator = new AtomicInteger(0);
     private final EventStore rawDataEventStore;
-    private final ZContext context = new ZContext(SystemProperties.getValueIgnoreCase(N_IO_THREADS, N_IO_THREADS_DEFAULT));
+    private final boolean owningContext;
+    private final ZContext context;
     private final ZMQ.Poller poller;
-    private final ZMQ.Socket controlSocket;
+    private final ZMQ.Socket sourceSocket;
     private final IoBuffer byteBuffer = new FastByteBuffer(0, true, null); // never actually used
     private final IoClassSerialiser ioClassSerialiser = new IoClassSerialiser(byteBuffer);
     private final String clientId;
     private final RbacProvider rbacProvider;
     private final ExecutorService executor; // NOPMD - threads are ok, not a webapp
-    private Thread thread;
-    private EventStore publicationTarget;
+    private final EventStore publicationTarget;
+    private AtomicReference<Thread> threadReference = new AtomicReference<>();
 
     static { // register default data sources
         DataSource.register(CmwLightDataSource.FACTORY);
@@ -121,25 +125,31 @@ public class DataSourcePublisher implements Runnable, Closeable {
     }
 
     public DataSourcePublisher(final RbacProvider rbacProvider, final ExecutorService executorService, final String... clientId) {
-        this.executor = executorService == null ? Executors.newCachedThreadPool() : executorService;
+        this(null, null, rbacProvider, executorService, clientId);
+        start(); // NOPMD
+    }
+
+    public DataSourcePublisher(final ZContext ctx, final EventStore publicationTarget, final RbacProvider rbacProvider, final ExecutorService executorService, final String... clientId) {
+        owningContext = ctx == null;
+        this.context = Objects.requireNonNullElse(ctx, new ZContext(SystemProperties.getValueIgnoreCase(N_IO_THREADS, N_IO_THREADS_DEFAULT)));
+        this.executor = Objects.requireNonNullElse(executorService, Executors.newCachedThreadPool());
         poller = context.createPoller(1);
         // control socket for adding subscriptions / triggering requests from other threads
-        controlSocket = context.createSocket(SocketType.DEALER);
-        controlSocket.bind(inprocCtrl);
-        poller.register(controlSocket, ZMQ.Poller.POLLIN);
-        // instantiate internal event store for raw messages
+        sourceSocket = context.createSocket(SocketType.DEALER);
+        sourceSocket.setHWM(SystemProperties.getValueIgnoreCase(HIGH_WATER_MARK, HIGH_WATER_MARK_DEFAULT));
+        sourceSocket.bind(inprocCtrl);
+        poller.register(sourceSocket, ZMQ.Poller.POLLIN);
+
+        // instantiate internal event store and handler for raw messages
+        // the internal handler deserialises updates, republishes them, or notifies external listeners/futures
         rawDataEventStore = EventStore.getFactory().setSingleProducer(true).setFilterConfig(EvtTypeFilter.class).build();
+        rawDataEventStore.register(this::internalEventHandler);
+
         // setup clientId to identify subscribers from server
         this.clientId = clientId.length == 1 ? clientId[0] : DataSourcePublisher.class.getName();
         this.rbacProvider = rbacProvider;
-        // setup internal event handler to deserialise updates and republish them or notify listeners/futures
-        rawDataEventStore.register(this::internalEventHandler);
-    }
 
-    public DataSourcePublisher(final EventStore publicationTarget, final RbacProvider rbacProvider, final ExecutorService executorService, final String... clientId) {
-        this(rbacProvider, executorService, clientId);
-        // register external event store
-        this.publicationTarget = publicationTarget;
+        this.publicationTarget = publicationTarget; // register external event store
     }
 
     public ZContext getContext() {
@@ -157,39 +167,38 @@ public class DataSourcePublisher implements Runnable, Closeable {
     @Override
     public void close() {
         shallRun.set(false);
+        Thread thread = threadReference.get();
         if (thread != null) {
             thread.interrupt();
             try {
                 thread.join(heartbeatInterval);
-            } catch (InterruptedException e) {
+            } catch (InterruptedException e) { // NOPMD NOSONAR - re-throwing with different type
                 throw new IllegalStateException(thread.getName() + " did not shut down in " + heartbeatInterval + " ms", e);
             }
         }
-        context.destroy();
+        thread = threadReference.get();
+        if (running.get() && thread != null) {
+            thread.interrupt();
+        }
+        if (owningContext) {
+            context.destroy();
+        }
     }
 
     public void start() {
-        final String threadName = "DataSourceProducerThread-" + clientId;
+        Thread thread = threadReference.get();
         if (thread != null) {
-            throw new IllegalStateException(threadName + " already running");
+            LOGGER.atWarn().addArgument(thread.getName()).log("Thread '{}' already running");
+            return;
         }
+        final String threadName = "DataSourceProducerThread-" + clientId;
         thread = new Thread(null, this, threadName, 0); // NOPMD - not a webapp
+        threadReference.set(thread);
         thread.setDaemon(true);
         thread.start();
     }
 
     public void stop() {
-        shallRun.set(false);
-        if (thread != null) {
-            try {
-                thread.join(heartbeatInterval);
-            } catch (InterruptedException e) {
-                throw new IllegalStateException(thread.getName() + " did not shut down in " + heartbeatInterval + " ms", e);
-            }
-            if (running.get()) {
-                thread.interrupt();
-            }
-        }
         close();
     }
 
@@ -210,6 +219,7 @@ public class DataSourcePublisher implements Runnable, Closeable {
                 dataAvailable = handleDataSourceSockets(); // get data from clients
                 dataAvailable |= handleControlSocket(); // check specifically for control socket
             }
+
             nextHousekeeping = clientMap.values().stream().mapToLong(DataSource::housekeeping).min().orElse(System.currentTimeMillis() + heartbeatInterval);
             timeOut = nextHousekeeping - System.currentTimeMillis();
         }
@@ -220,21 +230,23 @@ public class DataSourcePublisher implements Runnable, Closeable {
         }
         rawDataEventStore.stop();
         running.set(false);
+        threadReference.set(null);
     }
 
     protected boolean handleControlSocket() {
-        final ZMsg controlMsg = ZMsg.recvMsg(controlSocket, false);
+        final ZMsg controlMsg = ZMsg.recvMsg(sourceSocket, false);
         if (controlMsg == null) {
             return false; // no more data available on control socket
         }
+
         // msgType, requestId and endpoint have to be always present
         assert controlMsg.size() >= MIN_FRAMES_INTERNAL_MSG : "ignoring invalid message - message size: " + controlMsg.size();
 
         final Command msgType = Command.getCommand(controlMsg.pollFirst().getData());
         final String requestId = requireNonNull(controlMsg.pollFirst()).getString(UTF_8);
         final URI endpoint = URI.create(requireNonNull(controlMsg.pollFirst()).getString(UTF_8));
-        final byte[] data = controlMsg.isEmpty() ? EMPTY_BYTE_ARRAY : controlMsg.pollFirst().getData();
-        final byte[] rbacToken = controlMsg.isEmpty() ? EMPTY_BYTE_ARRAY : controlMsg.pollFirst().getData();
+        final byte[] data = controlMsg.isEmpty() ? EMPTY_FRAME : controlMsg.pollFirst().getData();
+        final byte[] rbacToken = controlMsg.isEmpty() ? EMPTY_FRAME : controlMsg.pollFirst().getData();
 
         final DataSource client = getClient(endpoint); // get client for endpoint
         switch (msgType) {
@@ -263,6 +275,7 @@ public class DataSourcePublisher implements Runnable, Closeable {
             if (reply == null || reply.isEmpty()) {
                 continue; // no data received, queue empty or there was data received, but only used for internal state of the client
             }
+
             dataAvailable = true;
             // the received data consists of the following frames:
             // replyType(byte), reqId(string), endpoint(string), dataBody(byte[])
@@ -282,29 +295,22 @@ public class DataSourcePublisher implements Runnable, Closeable {
                 event.payload.set(internalData); // InternalDomainObject containing - ZMsg containing header, body and exception frame + Future
 
                 final EvtTypeFilter evtTypeFilter = event.getFilter(EvtTypeFilter.class);
-                evtTypeFilter.updateType = returnFuture.replyType;
+                evtTypeFilter.updateType = returnFuture.requestType;
                 evtTypeFilter.property = URI.create(endpoint);
             });
         }
         return dataAvailable;
     }
 
-    protected <R> ThePromisedFuture<R, ?> newRequestFuture(final URI endpoint,
-            final Class<R> requestedDomainObjType,
-            final Command replyType,
-            final String requestId) {
+    protected <R> ThePromisedFuture<R, ?> newRequestFuture(final URI endpoint, final Class<R> requestedDomainObjType, final Command requestType, final String requestId) {
         FilterRegistry.checkClassForNewFilters(requestedDomainObjType);
-        final ThePromisedFuture<R, ?> requestFuture = new ThePromisedFuture<>(endpoint, requestedDomainObjType, null, replyType, requestId, null);
+        final ThePromisedFuture<R, ?> requestFuture = new ThePromisedFuture<>(endpoint, requestedDomainObjType, null, requestType, requestId, null);
         final Object oldEntry = requests.put(requestId, requestFuture);
         assert oldEntry == null : "requestID '" + requestId + "' already present in requestFutureMap";
         return requestFuture;
     }
 
-    protected <R, C> ThePromisedFuture<R, C> newSubscriptionFuture(final URI endpoint,
-            final Class<R> requestedDomainObjType,
-            final Class<C> contextType,
-            final String requestId,
-            final NotificationListener<R, C> listener) {
+    protected <R, C> ThePromisedFuture<R, C> newSubscriptionFuture(final URI endpoint, final Class<R> requestedDomainObjType, final Class<C> contextType, final String requestId, final NotificationListener<R, C> listener) {
         FilterRegistry.checkClassForNewFilters(requestedDomainObjType);
         final ThePromisedFuture<R, C> requestFuture = new ThePromisedFuture<>(endpoint, requestedDomainObjType, contextType, Command.SUBSCRIBE, requestId, listener);
         final Object oldEntry = requests.put(requestId, requestFuture);
@@ -337,9 +343,13 @@ public class DataSourcePublisher implements Runnable, Closeable {
         Object replyDomainObject = null;
         if (exceptionMsg.isBlank()) {
             try {
-                if (body.length != 0) {
+                final Class<?> reqClassType = domainObject.future.getRequestedDomainObjType();
+                if (reqClassType.isAssignableFrom(BinaryData.class)) {
+                    // special case: handle byte[] and/or -- albeit UTF8 byte[] encoded -- String data
+                    replyDomainObject = new BinaryData(domainObject.future.getInternalRequestID(), MimeType.BINARY, body);
+                } else if (body.length != 0) {
                     ioClassSerialiser.setDataBuffer(FastByteBuffer.wrap(body));
-                    replyDomainObject = ioClassSerialiser.deserialiseObject(domainObject.future.getRequestedDomainObjType());
+                    replyDomainObject = ioClassSerialiser.deserialiseObject(reqClassType);
                     ioClassSerialiser.setDataBuffer(byteBuffer); // allow received byte array to be released
                 }
                 if (notifyFuture) {
@@ -416,13 +426,13 @@ public class DataSourcePublisher implements Runnable, Closeable {
     }
 
     public class Client implements Closeable {
-        private final ZMQ.Socket controlSocket;
+        private final ZMQ.Socket clientSocket;
         private IoBuffer byteBuffer;
         private IoClassSerialiser ioClassSerialiser;
 
         private Client() { // accessed via outer class method
-            controlSocket = context.createSocket(SocketType.DEALER);
-            controlSocket.connect(inprocCtrl);
+            clientSocket = context.createSocket(SocketType.DEALER);
+            clientSocket.connect(inprocCtrl);
         }
 
         public <R, C> Future<R> get(URI endpoint, final C requestContext, final Class<R> requestedDomainObjType, final RbacProvider... rbacProvider) {
@@ -459,12 +469,12 @@ public class DataSourcePublisher implements Runnable, Closeable {
             msg.add(Command.UNSUBSCRIBE.getData());
             msg.add(requestId);
             msg.add(requests.get(requestId).endpoint.toString());
-            msg.send(controlSocket);
+            msg.send(clientSocket);
         }
 
         @Override
         public void close() {
-            controlSocket.close();
+            clientSocket.close();
         }
 
         private IoClassSerialiser getSerialiser() {
@@ -488,15 +498,15 @@ public class DataSourcePublisher implements Runnable, Closeable {
             return endpointQuery;
         }
 
-        private <R, C> void request(final String requestId, final Command replyType, final URI endpoint, R requestBody, C requestContext, final RbacProvider... rbacProvider) {
+        private <R, C> void request(final String requestId, final Command requestType, final URI endpoint, R requestBody, C requestContext, final RbacProvider... rbacProvider) {
             FilterRegistry.checkClassForNewFilters(requestContext);
             // signal socket for get with endpoint and request id
             final ZMsg msg = new ZMsg();
-            msg.add(replyType.getData());
+            msg.add(requestType.getData());
             msg.add(requestId);
             msg.add(endpoint.toString());
             if (requestBody == null) {
-                msg.add(EMPTY_FRAME);
+                msg.add(EMPTY_ZFRAME);
             } else {
                 final Class<? extends IoSerialiser> matchingSerialiser = DataSource.getFactory(endpoint).getMatchingSerialiserType(endpoint);
                 final IoClassSerialiser serialiser = getSerialiser(); // lazily initialize IoClassSerialiser
@@ -510,9 +520,12 @@ public class DataSourcePublisher implements Runnable, Closeable {
                 // rbac.sign(msg); // todo: sign message and add rbac token and signature
                 LOGGER.atWarn().log("RbacProvider not yet implemented");
             } else {
-                msg.add(EMPTY_FRAME);
+                msg.add(EMPTY_ZFRAME);
             }
-            msg.send(controlSocket);
+
+            if (!msg.send(clientSocket)) {
+                LOGGER.atWarn().addArgument(requestType).addArgument(msg).log("could not send {} from client to source - message: {}");
+            }
         }
     }
 
@@ -520,21 +533,16 @@ public class DataSourcePublisher implements Runnable, Closeable {
         private final URI endpoint;
         private final Class<R> requestedDomainObjType;
         private final Class<C> contextType;
-        private final Command replyType;
+        private final Command requestType;
         private final String internalRequestID;
         private final NotificationListener<R, C> listener;
 
-        public ThePromisedFuture(final URI endpoint,
-                final Class<R> requestedDomainObjType,
-                final Class<C> contextType,
-                final Command replyType,
-                final String internalRequestID,
-                final NotificationListener<R, C> listener) {
+        public ThePromisedFuture(final URI endpoint, final Class<R> requestedDomainObjType, final Class<C> contextType, final Command requestType, final String internalRequestID, final NotificationListener<R, C> listener) {
             super();
             this.endpoint = endpoint;
             this.requestedDomainObjType = requestedDomainObjType;
             this.contextType = contextType;
-            this.replyType = replyType;
+            this.requestType = requestType;
             this.internalRequestID = internalRequestID;
             this.listener = listener;
         }
@@ -544,7 +552,7 @@ public class DataSourcePublisher implements Runnable, Closeable {
         }
 
         public Command getReplyType() {
-            return replyType;
+            return requestType;
         }
 
         public Class<R> getRequestedDomainObjType() {
@@ -570,7 +578,7 @@ public class DataSourcePublisher implements Runnable, Closeable {
         }
     }
 
-    public static class InternalDomainObject { //TODO: unclean design -- should not export this to external clients
+    protected static class InternalDomainObject {
         public final ZMsg data;
         public final ThePromisedFuture<?, ?> future;
